Shifted parseExpression function to func/parseExpression/parseExpression.cpp
